Refactor PerformanceOptimizationService
Refactor `src/services/PerformanceOptimizationService.ts` into `PerformanceProfileManager.ts`, `OptimizationApplier.ts`, and `PerformanceMonitor.ts`, updating the original `PerformanceOptimizationService.ts` to orchestrate these new components. All functionality preserved.
